\name{Anduki}
\docType{data}
\alias{Anduki}
\title{Anduki Forest Reserve}
\description{
  A set of coordinates delimiting the Anduki Forest Reserve in Brunei.
}
\usage{data(Anduki)}
\format{
  A two-column matrix giving the coordinates in degrees (longitude and
  latitude, respectively).
}
\source{
  WDPA: World Database of Protected Areas. April 2024.
}
\keyword{datasets}
