\name{tigers-package}
\alias{tigers-package}
\alias{tigers}
\docType{package}
\title{Integration of Geography, Environment, and Remote Sensing}
\description{
  \pkg{tigers} provides functions for manipulating polygons,
  coordinates, \dots

  All the tools programmed in \pkg{tigers} are ``class-free'': they work
  on numeric vectors or matrices (even data frames) that store
  coordinates. So the functions in the present package can easily be
  interfaced with other packages such as \pkg{terra}, \pkg{sf}, or
  \pkg{sp}.

  The majority of the computations done by \pkg{tigers} are performed by
  efficient C code which could be interfaced with other languages (e.g.,
  Python).

  The complete list of functions can be displayed with
  \code{library(help = tigers)}.
}

\author{
  Emmanuel Paradis

  Maintainer: Emmanuel Paradis <Emmanuel.Paradis@ird.fr>
}
\keyword{package}
