\name{tigers-package}
\alias{tigers-package}
\alias{tigers}
\docType{package}
\title{Tools for Integrating Geography, Environment, and Remote Sensing}
\description{
  \pkg{tigers} provides functions for manipulating polygons,
  coordinates, \dots

  The complete list of functions can be displayed with
  \code{library(help = tigers)}.
}

\author{
  Emmanuel Paradis

  Maintainer: Emmanuel Paradis <Emmanuel.Paradis@ird.fr>
}
\keyword{package}
